Add images as resources for various buttons.
Use optXxx.isEmpty() instead of !optXxx.isPresent().

Rework ImageAppearanceControlPanel to use the new GenericPanelManager class in Pipestone.

Rework the media library browser and the image viewer classes to allow them to share more code.

Lots of cleanup and some reworking of the node flavour and node type classes (including renaming them to InfoNodeFlavour and InfoNodeType).

Various classes to manage Lancot data entities like references and statements and such.

A fair bit of refactoring in a wide variety of classes.